modularize multinomial
